+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-## [alpha-0.7.11] - 
+## [alpha-0.7.11] - 2025.04.14
 
 **Fixes**
 - Double trouble display correctly now.
-- Fix to recover the current ongoing seasonal event correctly.
+- Fix to recover the current ongoing seasonal event on new Log files.
 
 ## [alpha-0.7.10] - 2025.04.13